port vtu_output from future
